TS isolatedModules: true, рефакторинг
