+++ src/app/providers/StoreProvider/index.ts
@@ -5,8 +5,11 @@ import { StoreProvider } from './ui/StoreProvider';
 export {
     StoreProvider,
     createReduxStore,
+};
+
+export type {
     StateSchema,
     ReduxStoreWithManager,
-    AppDispatch,
     ThunkConfig,
+    AppDispatch,
 };
