+++ src/entities/Article/model/consts/articleConsts.ts
@@ -1,0 +1,23 @@
+export enum ArticleSortField {
+    VIEWS = 'views',
+    CREATED = 'createdAt',
+    TITLE = 'title',
+}
+
+export enum ArticleBlockTypes {
+    CODE = 'CODE',
+    TEXT = 'TEXT',
+    IMAGE = 'IMAGE',
+}
+
+export enum ArticleTypes {
+    ALL = 'ALL',
+    IT = 'IT',
+    SCIENCE = 'SCIENCE',
+    ECONOMICS = 'ECONOMICS'
+}
+
+export enum ArticleView {
+    LIST = 'LIST',
+    GRID = 'GRID'
+}
