+++ src/entities/Article/model/types/article.ts
@@ -1,16 +1,5 @@
 import { User } from 'entities/User';
-
-export enum ArticleSortField {
-    VIEWS = 'views',
-    CREATED = 'createdAt',
-    TITLE = 'title',
-}
-
-export enum ArticleBlockTypes {
-    CODE = 'CODE',
-    TEXT = 'TEXT',
-    IMAGE = 'IMAGE',
-}
+import { ArticleBlockTypes, ArticleTypes } from '../consts/articleConsts';
 
 export interface ArticleBlockBase {
     id: string;
@@ -35,18 +24,6 @@ export interface ArticleImageBlock extends ArticleBlockBase{
 }
 
 export type ArticleBlocks = ArticleCodeBlock | ArticleTextBlock | ArticleImageBlock;
-
-export enum ArticleTypes {
-    ALL = 'ALL',
-    IT = 'IT',
-    SCIENCE = 'SCIENCE',
-    ECONOMICS = 'ECONOMICS'
-}
-
-export enum ArticleView {
-    LIST = 'LIST',
-    GRID = 'GRID'
-}
 
 export interface Article {
     id: string;
