+++ src/entities/Article/ui/ArticleDetails/ArticleDetails.stories.tsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
 import { StoreDecorator } from 'shared/config/storybook/StoreDecorator/StoreDecorator';
+import { ArticleBlockTypes, ArticleTypes } from '../../model/consts/articleConsts';
 import { ArticleDetails } from './ArticleDetails';
-import { Article, ArticleBlockTypes, ArticleTypes } from '../../model/types/article';
+import { Article } from '../../model/types/article';
 
 export default {
     title: 'entities/Article/ArticleDetails',
