+++ src/entities/Article/ui/ArticleSortSelector/ArticleSortSelector.tsx
@@ -3,8 +3,8 @@ import { memo, useMemo } from 'react';
 import { Select, SelectOptions } from 'shared/ui/Select/Select';
 import { useTranslation } from 'react-i18next';
 import { SortOrder } from 'shared/types/sort';
+import { ArticleSortField } from '../../model/consts/articleConsts';
 import cls from './ArticleSortSelector.module.scss';
-import { ArticleSortField } from '../../model/types/article';
 
 interface ArticleSortSelectorProps {
     className?: string;
