+++ src/entities/Article/ui/ArticleTypeTabs/ArticleTypeTabs.tsx
@@ -2,7 +2,7 @@ import { classNames } from 'shared/lib/classNames/classNames';
 import { memo, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { TabItem, Tabs } from 'shared/ui/Tabs/Tabs';
-import { ArticleTypes } from '../../model/types/article';
+import { ArticleTypes } from '../../model/consts/articleConsts';
 
 interface ArticleTypeTabsProps {
     className?: string;
