+++ src/entities/Article/ui/ArticleViewSelector/ArticleViewSelector.tsx
@@ -5,7 +5,7 @@ import ListIcon from 'shared/assets/icons/viewList-24-24.svg';
 import { Button, ButtonTheme } from 'shared/ui/Button/Button';
 import { Icon } from 'shared/ui/Icon/Icon';
 import cls from './ArticleViewSelector.module.scss';
-import { ArticleView } from '../../model/types/article';
+import { ArticleView } from '../../model/consts/articleConsts';
 
 interface ArticleViewSelectorProps {
     className?: string;
