+++ src/entities/Profile/index.ts
@@ -1,7 +1,10 @@
-import { Profile } from './model/types/profile';
+import type { Profile } from './model/types/profile';
 import { ProfileCard } from './ui/ProfileCard/ProfileCard';
 
 export {
-    Profile,
     ProfileCard,
+};
+
+export type {
+    Profile,
 };
