+++ src/entities/User/index.ts
@@ -1,20 +1,24 @@
+import { UserRoles } from './model/consts/userConsts';
 import { getAuthUserData } from './model/selectors/getAuthUserData/getAuthUserData';
 import { getUserInited } from './model/selectors/getUserInited/getUserInited';
 import { getUserRoles } from './model/selectors/getUserRoles/getUserRoles';
 import { isUserAdmin } from './model/selectors/isUserAdmin/isUserAdmin';
 import { isUserManager } from './model/selectors/isUserManager/isUserManager';
 import { userActions, userReducer } from './model/slice/userSlice';
-import { User, UserRoles, UserSchema } from './model/types/userSchema';
+import type { User, UserSchema } from './model/types/userSchema';
 
 export {
     userReducer,
     userActions,
-    UserSchema,
-    User,
     getAuthUserData,
     getUserInited,
     isUserAdmin,
     isUserManager,
     UserRoles,
     getUserRoles,
+};
+
+export type {
+    UserSchema,
+    User,
 };
