+++ src/entities/User/model/consts/userConsts.ts
@@ -1,0 +1,5 @@
+export enum UserRoles {
+    ADMIN = 'ADMIN',
+    USER = 'USER',
+    MANAGER = 'MANAGER',
+}
