+++ src/entities/User/model/selectors/isUserManager/isUserManager.ts
@@ -1,5 +1,5 @@
 import { createSelector } from '@reduxjs/toolkit';
+import { UserRoles } from '../../consts/userConsts';
 import { getUserRoles } from '../getUserRoles/getUserRoles';
-import { UserRoles } from '../../types/userSchema';
 
 export const isUserManager = createSelector(getUserRoles, (roles) => roles?.includes(UserRoles.MANAGER));
