+++ src/features/AddCommentForm/index.ts
@@ -1,7 +1,10 @@
-import { AddCommentFormSchema } from './model/types/addCommentForm';
+import type { AddCommentFormSchema } from './model/types/addCommentForm';
 import { AddCommentFormAsync as AddCommentForm } from './ui/AddCommentForm/AddCommentForm.async';
 
 export {
-    AddCommentFormSchema,
     AddCommentForm,
+};
+
+export type {
+    AddCommentFormSchema,
 };
