+++ src/features/AuthByUsername/index.ts
@@ -1,7 +1,10 @@
-import { LoginSchema } from './model/types/loginSchema';
+import type { LoginSchema } from './model/types/loginSchema';
 import { LoginModal } from './ui/LoginModal/LoginModal';
 
 export {
     LoginModal,
+};
+
+export type {
     LoginSchema,
 };
