+++ src/features/EditableProfileCard/model/services/updateProfileData/updateProfileData.test.ts
@@ -1,8 +1,8 @@
 import { TestAsyncThunk } from 'shared/lib/tests/TestAsyncThunk/TestAsyncThunk';
 import { Currency } from 'entities/Currency';
 import { Country } from 'entities/Country';
+import { ValidateProfileErrors } from '../../consts/editableProfileCardConsts';
 import { updateProfileData } from './updateProfileData';
-import { ValidateProfileErrors } from '../../types/editableProfileCardSchema';
 
 const data = {
     first: 'Tom',
