+++ src/features/EditableProfileCard/model/services/validateProfileData/validateProfileData.test.ts
@@ -1,7 +1,7 @@
 import { Currency } from 'entities/Currency';
 import { Country } from 'entities/Country';
+import { ValidateProfileErrors } from '../../consts/editableProfileCardConsts';
 import { validateProfileData } from './validateProfileData';
-import { ValidateProfileErrors } from '../../types/editableProfileCardSchema';
 
 const data = {
     first: 'Tom',
