+++ src/features/EditableProfileCard/model/types/editableProfileCardSchema.ts
@@ -1,17 +1,5 @@
 import { Profile } from 'entities/Profile';
-
-export enum ValidateProfileErrors {
-    INCORRECT_FIRST = 'INCORRECT_FIRST',
-    INCORRECT_LASTNAME = 'INCORRECT_LASTNAME',
-    INCORRECT_AGE = 'INCORRECT_AGE',
-    INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
-    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
-    INCORRECT_CITY = 'INCORRECT_CITY',
-    INCORRECT_USERNAME = 'INCORRECT_USERNAME',
-    INCORRECT_AVATAR = 'INCORRECT_AVATAR',
-    NO_DATA = 'NO_DATA',
-    SERVER_ERROR = 'SERVER_ERROR',
-}
+import { ValidateProfileErrors } from '../consts/editableProfileCardConsts';
 
 export interface ProfileSchema {
     data?: Profile;
