+++ src/features/ScrollRestoration/index.ts
@@ -1,10 +1,13 @@
 import { getScrollRestorationByPath } from './model/selectors/getScrollRestorationByPath/getScrollRestorationByPath';
 import { scrollRestorationActions, scrollRestorationReducer } from './model/slice/scrollRestorationSlice';
-import { ScrollRestorationSchema } from './model/types/ScrollRestorationSchema';
+import type { ScrollRestorationSchema } from './model/types/ScrollRestorationSchema';
 
 export {
-    ScrollRestorationSchema,
     scrollRestorationReducer,
     scrollRestorationActions,
     getScrollRestorationByPath,
+};
+
+export type {
+    ScrollRestorationSchema,
 };
