+++ src/pages/ArticlesPage/index.ts
@@ -1,7 +1,10 @@
-import { ArticlesPageSchema } from './model/types/articlesPageSchema';
+import type { ArticlesPageSchema } from './model/types/articlesPageSchema';
 import { ArticlesPageAsync as ArticlesPage } from './ui/ArticlesPage/ArticlesPage.async';
 
 export {
-    ArticlesPageSchema,
     ArticlesPage,
+};
+
+export type {
+    ArticlesPageSchema,
 };
