+++ src/pages/ArticlesPage/ui/ArticlesPageFilters/ArticlesPageFilters.tsx
@@ -2,7 +2,7 @@ import { classNames } from 'shared/lib/classNames/classNames';
 import { memo, useCallback } from 'react';
 import {
     ArticleSortField,
-    ArticleSortSelector,
+    ArticleSortSelector, ArticleTypes,
     ArticleTypeTabs, ArticleView,
     ArticleViewSelector,
 } from 'entities/Article';
@@ -13,7 +13,6 @@ import { Card } from 'shared/ui/Card/Card';
 import { Input } from 'shared/ui/Input/Input';
 import { SortOrder } from 'shared/types/sort';
 import { useDebounce } from 'shared/lib/hooks/useDebounce/useDebounce';
-import { ArticleTypes } from 'entities/Article/model/types/article';
 import { fetchArticlesList } from '../../model/services/fetchArticlesList/fetchArticlesList';
 import { articlesPageActions } from '../../model/slices/articlesPageSlice';
 import cls from './ArticlesPageFilters.module.scss';
