+++ tsconfig.json
@@ -13,7 +13,8 @@
       "*": ["./src/*"]
     },
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "isolatedModules": true,
   },
   "ts-node": {
     "compilerOptions": {
